Refactor API clients and models
