+++ app/models/index.js
@@ -1,0 +1,7 @@
+module.exports = Object.assign(
+  {},
+  require('./base'),
+  require('./control_panel_api'),
+  require('./kubernetes'),
+  require('./github')
+);
