+++ test/test-k8s-models.js
@@ -1,8 +1,7 @@
 const { assert } = require('chai');
 const { mock_api } = require('./conftest');
-const { api } = require('../app/k8s-api-client');
-const { ModelSet } = require('../app/base-model');
-const { Pod, Deployment } = require('../app/models');
+const { api } = require('../app/api_clients/kubernetes');
+const { Deployment, ModelSet, Pod } = require('../app/models');
 const deployments = require('./fixtures/deployments').items;
 const pods_response = require('./fixtures/deployment-pods');
